# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: LeadForge AI REST API
  version: 1.0.0
  description: |
    Complete REST API for LeadForge AI.
    All endpoints are prefixed with `/api/v1`.
    Two auth modes:
    1. **Clerk JWT** — short-lived Bearer token from the Clerk session. Optional `X-Team-Id` header selects active team; defaults to first membership.
    2. **API Key** — long-lived `lf_*` token (Bearer). Bound to a single team. Listed/managed at `/api-keys`.
servers:
  - url: https://api.leadforge.ai/api/v1
  - url: http://localhost:4000/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or lf_api_key
  parameters:
...
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
  schemas:
...
    UserRole: { type: string, enum: [OWNER, ADMIN, MEMBER, VIEWER] }
...
    VerificationStatus: { type: string, enum: [PENDING, VALID, INVALID, RISKY, UNKNOWN, CATCH_ALL] }
...
    FilterCondition:
      type: object
      required: [field, operator, value]
      properties:
        field:
          type: string
          enum: [niche, industry, domain_extension, country, region, city, language, keyword, companyName, companyDomain, companySize, companyRevenue, companyAuthority, jobTitle, jobRole, jobSeniority, jobDepartment, technology, social_presence, has_email, has_phone, has_contact_page, qualityScore, relevanceScore, intentScore, authorityScore, verificationStatus, createdAt, sourceType]
        operator:
          type: string
          enum: [eq, neq, in, nin, contains, starts_with, ends_with, gt, gte, lt, lte, between, exists, not_exists, has_any, has_all]
        value: {}
    Filter:
      oneOf:
        - $ref: '#/components/schemas/FilterCondition'
        - type: object
          properties:
            AND: { type: array, items: { $ref: '#/components/schemas/Filter' } }
        - type: object
          properties:
            OR: { type: array, items: { $ref: '#/components/schemas/Filter' } }
        - type: object
          properties:
            NOT: { $ref: '#/components/schemas/Filter' }
...
    EmailVerification:
      type: object
      properties:
        email: { type: string }
        status: { $ref: '#/components/schemas/VerificationStatus' }
        score: { type: integer }
        confidence: { type: number }
        syntaxValid: { type: boolean }
        isDisposable: { type: boolean }
        mxValid: { type: boolean }
        smtpDeliverable: { type: boolean, nullable: true }
        isCatchAll: { type: boolean, nullable: true }
        isRoleAccount: { type: boolean }
        reason: { type: string }
paths:
...
  /teams:
    get: { summary: List my teams, responses: { '200': { description: ok } } }
    post:
      summary: Create team
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, properties: { name: { type: string } }, required: [name] }
      responses: { '201': { description: created } }
  /teams/{id}:
    get: { summary: Get team, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
    patch: { summary: Update team, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
  /teams/{id}/invitations:
    post:
      summary: Invite member
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { $ref: '#/components/schemas/UserRole' }
      responses: { '201': { description: created } }
...
  /leads:
    get:
      summary: List leads
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: jobId, in: query, schema: { type: string } }
        - { name: verificationStatus, in: query, schema: { $ref: '#/components/schemas/VerificationStatus' } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, qualityScore, companyName, fullName] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses: { '200': { description: ok } }
  /leads/search:
    post:
      summary: Search leads via filter tree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filter]
              properties:
                filter: { $ref: '#/components/schemas/Filter' }
                page: { type: integer }
                pageSize: { type: integer }
      responses: { '200': { description: ok } }
  /leads/{id}:
    get: { summary: Get lead, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: ok } } }
    patch:
      summary: Update lead
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [NEW, VERIFIED, ENRICHED, EXPORTED, ARCHIVED] }
                isFavorite: { type: boolean }
                notes: { type: string, nullable: true }
      responses: { '200': { description: ok } }
    delete: { summary: Delete lead, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '204': { description: deleted } } }
  /leads/bulk:
    post:
      summary: Bulk action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, action]
              properties:
                ids: { type: array, items: { type: string } }
                action: { type: string, enum: [archive, favorite, unfavorite, delete, verify] }
      responses: { '200': { description: ok } }
  /verification/single:
    post:
      summary: Verify one email (synchronous)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email], properties: { email: { type: string, format: email } } }
      responses:
        '200':
          description: ok
          content: { application/json: { schema: { type: object, properties: { result: { $ref: '#/components/schemas/EmailVerification' } } } } }
  /verification/batch:
    post:
      summary: Verify many leads (async)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [leadIds]
              properties:
                leadIds: { type: array, items: { type: string } }
      responses: { '202': { description: queued } }
  /verification/status/{leadId}:
    get:
      summary: Verification result for a lead
      parameters: [{ name: leadId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
...
  /webhooks:
    post:
      summary: Create webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri }
                events: { type: array, items: { type: string, enum: [job.completed, job.failed, lead.created, verification.completed, export.ready] } }
      responses: { '201': { description: created } }
    get: { summary: List webhooks, responses: { '200': { description: ok } } }
  /webhooks/{id}:
    delete: { summary: Delete webhook, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { '204': { description: deleted } } }
...